Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 182ms, Memory: 10MB, Status: Passed - CodeTree

+++ 241029/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -16,7 +16,7 @@ public class Main {
         a = sc.nextLong();
         b = sc.nextLong();
 
-        for (int target = a; target <= b; target ++) {
+        for (long target = a; target <= b; target ++) {
             long count = bisearch(target);
             if (count < minCount) {
                 minCount = count;
